// src/interfaces/workingHistory.ts
export interface IworkingHistory {
    Id: number;
    Employee_ID_Number: number;
    Company_Name: string;
    Employer_Name: string;
    Company_Employer_Address: string;
    Company_Employer_Cellular_Phone: number;
    Company_Employer_Office_Phone: number;
    Previous_Qualification: string;
    Previous_Experience: string;
    p_Start_Date_Day: number;
    p_Start_Date_Month: number;
    p_Start_Date_Year: number;
    p_End_Date_Day: number;
    p_End_Date_Month: number;
    p_End_Date_Year: number;
}